+++ liana/tests/test_causalnet.py
@@ -20,14 +20,12 @@ def test_build_prior_network():
     
 
 def test_caulsalnet():
-    prior_graph = build_prior_network(input_pkn, input_scores, output_scores, verbose=True)
+    prior_graph = build_prior_network(input_pkn, input_scores, output_scores, verbose=False)
     df_res, problem = find_causalnet(prior_graph, 
                                      input_scores, 
                                      output_scores, 
                                      node_weights=node_weights,
-                                     verbose=False,
-                                     show_solver_output=False,
-                                     solver='scipy'
+                                     verbose=False
                                      )
 
     assert problem.weights == [1.0, 0.01, 1.0]
@@ -35,5 +33,16 @@ def test_caulsalnet():
     assert df_res['target_pred_val'].values.sum() == 8
     assert df_res[df_res['source_type']=='input']['source'].values[0] == 'I2'
     assert (df_res[df_res['target_type']=='output']['target'].isin(['M1', 'M2'])).all()
+    
 
 
+def test_causalnet_noweights():
+    prior_graph = build_prior_network(input_pkn, input_scores, output_scores, verbose=False)
+    df_res, problem = find_causalnet(prior_graph,
+                                     input_scores,
+                                     output_scores,
+                                     node_weights={"N1": 1, "N2": 0},
+                                     verbose=False
+                                     )
+    assert df_res['source_pred_val'].values.sum() == 9
+    assert df_res['target_pred_val'].values.sum() == 10