+++ docs/source/release_notes.rst
@@ -1,6 +1,17 @@
 Release notes
 =============
 
+1.0.2 (05.10.2023)
+-------------------------------------------------
+- Added as `seed` param to `find_causalnet`, used to a small amount of noise to the nodes in to avoid obtaining multiple solutions to the same problem when multiple equal solutions are possible.
+
+- Updated `installation.rst`
+
+
+1.0.1 Stable Release (30.09.2023)
+-------------------------------------------------
+- Bumped CORNETO version and it's now installed via PyPI.
+
 1.0.0a2 (19.09.2023)
 
 - Interactions names in `tileplot` and `dotplot` will now be sorted according to `orderby` when used; related to #55